build(diagrams): esm build
