+++ packages/diagrams/package.json
@@ -20,6 +20,7 @@
   "types": "src/index.ts",
   "exports": {
     ".": "./src/index.ts",
+    "./responsive": "./src/responsive/index.ts",
     "./src": "./src/index.ts",
     "./src/diagram": "./src/diagram/index.ts",
     "./src/*": [
@@ -34,15 +35,16 @@
   "publishConfig": {
     "registry": "https://registry.npmjs.org",
     "access": "public",
-    "main": "dist/index.cjs",
     "module": "dist/index.js",
     "types": "dist/index.d.ts",
     "exports": {
       ".": {
         "types": "./dist/index.d.ts",
-        "module": "./dist/index.js",
-        "import": "./dist/index.js",
-        "require": "./dist/index.cjs"
+        "import": "./dist/index.js"
+      },
+      "./responsive": {
+        "types": "./dist/responsive/index.d.ts",
+        "import": "./dist/responsive/index.js"
       },
       "./src": "./src/index.ts",
       "./src/diagram": "./src/diagram/index.ts",
@@ -91,6 +93,7 @@
     "typescript": "^5.1.6",
     "unocss": "^0.54.1",
     "vite": "^4.4.8",
+    "vite-plugin-dts": "^3.5.1",
     "vite-plugin-watch-and-run": "^1.1.3"
   },
   "volta": {
