+++ packages/diagrams/src/diagram/index.ts
@@ -1,4 +1,2 @@
 export { Diagram } from './Diagram'
-export { LikeC4Diagram } from './LikeC4Diagram'
-export type { DiagramProps, DiagramApi } from './Diagram'
 export type * from './types'
