+++ packages/diagrams/src/diagram/shapes/index.ts
@@ -1,7 +1,8 @@
-export * from './Compound'
-export * from './Rectangle'
-export * from './Cylinder'
-export * from './Queue'
-export * from './nodeShape'
-export * from './nodeSprings'
-export * from './Edge'
+export { CompoundShape } from './Compound'
+export { RectangleShape } from './Rectangle'
+export { CylinderShape } from './Cylinder'
+export { MobileShape } from './Mobile'
+export { QueueShape } from './Queue'
+export { PersonShape } from './Person'
+export { nodeShape } from './nodeShape'
+export { EdgeShape } from './Edge'
