+++ packages/diagrams/src/diagram/shapes/utils.ts
@@ -1,13 +1,13 @@
-import type { OnPointerEvent } from './types'
+import type { KonvaPointerEvent } from '../types'
 
-export const mousePointer = (e: OnPointerEvent) => {
+export const mousePointer = (e: KonvaPointerEvent) => {
   const container = e.target.getStage()?.container()
   if (container) {
     container.style.cursor = 'pointer'
   }
 }
 
-export const mouseDefault = (e: OnPointerEvent) => {
+export const mouseDefault = (e: KonvaPointerEvent) => {
   const container = e.target.getStage()?.container()
   if (container) {
     container.style.cursor = 'auto'
