--- packages/diagrams/src/diagram/theme.ts
@@ -1,5 +1,0 @@
-import { defaultTheme } from '@likec4/core/colors'
-
-const DefaultDiagramTheme = defaultTheme
-
-export default DefaultDiagramTheme
