+++ packages/diagrams/src/responsive/index.ts
@@ -1,0 +1,3 @@
+export { ResponsiveDiagram } from './ResponsiveDiagram'
+
+export type { ResponsiveDiagramProps, DiagramApi } from './ResponsiveDiagram'
