+++ packages/tsconfig/base.json
@@ -3,9 +3,9 @@
   "display": "Base Strictest",
   "compilerOptions": {
     "lib": ["ESNext"],
-    "module": "es2022",
-    "target": "es2022",
-    "moduleResolution": "bundler",
+    "module": "ES2022",
+    "target": "ES2022",
+    "moduleResolution": "Bundler",
 
     "declaration": true,
     "declarationMap": false,
